Use `figcaption` directly for citation instead of `cite`
The use of `cite` as the body of the citation is called out as a behavior
to avoid according to the HTML5 specification. It's purpose is to cite
specific titles or references, but not for the generic citation itself.

In this patch we're using `figcaption` itself as exemplified in the
HTML5 spec. Any use of `cite` should probably come through a text
format within `RichText`, not as a part of the quote representation.